Rename SerialTaskQueue.hpp -> TaskQueue.hpp
